+++ grunt.tasks.js
@@ -11,7 +11,7 @@ module.exports = function(grunt) {
         'webpack',
         'inline',
         'htmlmin',
-        'string-replace:manifest-html',
+        'string-replace:service-worker',
         'string-replace:manifest',
         'copy:dist-icons',
         'copy:dist-manifest'
@@ -19,8 +19,7 @@ module.exports = function(grunt) {
 
     grunt.registerTask('build-desktop-app-content', [
         'copy:desktop-app-content',
-        'string-replace:desktop-public-key',
-        'string-replace:desktop-html'
+        'string-replace:desktop-public-key'
     ]);
 
     grunt.registerTask('build-desktop-update', [