+++ release-notes.md
@@ -9,6 +9,7 @@ Release notes
 `+` #348: configurable system-wide shortcuts  
 `+` #743: copying entry fields to clipboard  
 `+` #713: markdown notes  
+`+` #336: moving entries across files  
 `*` devtools are now opened with alt-cmd-I  
 `-` fix #764: multiple attachments display  
 `-` fix multi-line fields display in history