+++ release-notes.md
@@ -11,6 +11,7 @@ Release notes
 `+` #713: markdown notes  
 `+` #336: moving entries across files  
 `*` #156: using ServiceWorker instead of AppCache  
+`+` #564: Steam OTP support  
 `*` devtools are now opened with alt-cmd-I  
 `-` fix #764: multiple attachments display  
 `-` fix multi-line fields display in history