+++ release-notes.md
@@ -10,6 +10,7 @@ Release notes
 `+` #743: copying entry fields to clipboard  
 `+` #713: markdown notes  
 `+` #336: moving entries across files  
+`+` #498: CSV import  
 `*` #156: using ServiceWorker instead of AppCache  
 `+` #564: Steam OTP support  
 `*` devtools are now opened with alt-cmd-I