+++ release-notes.md
@@ -20,6 +20,7 @@ Release notes
 `-` fix #554: checking active window id during auto-type  
 `-` fix plugin gallery layout on mobile  
 `-` fix #1141: opening file from storage if cache is not available  
+`-` fix a color flash on start  
 
 ##### v1.10.1 (2019-09-20)
 `-` fixed Argon2 issues in Safari 13